+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixed
 * `modname_to_modpath` now handles cases where editable packages have modules where the name is different than the package.
+* Update `xdoctest.plugin` to support pytest 8.0
 
 
 ## Version 1.1.2 - Released 2023-010-25